Split the index entry if the user passed in a really big write
This is an optimization for checksum. Whenever we want to read
some data from a checksum enabled plfs file, then the whole
index entry needs to be read so that we can verify the checksum.
However when the index entry is very large(GBs), this will greatly
slow down the read operation.

So split the index entries at the write operation to avoid this.